pyproject.toml
[tool.poetry]
name = "extractor"
version = "0.1.0"
description = "Package create for using LLM's for extracting information from documents in a structured way."
authors = ["Pedro Cunha <pedro.cunha@asq.capital>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = ">=0.0.330"
tqdm = ">=4.66.1"
instructor = ">=0.2.9"
toml = ">=0.10.2"
invoke = ">=2.2.0"
pymupdf = ">=1.23.5"
loguru = ">=0.7.2"
python-dotenv = ">=1.0.0"
tiktoken = ">=0.5.1"
pandas = ">=2.1.2"
marvin = "^1.5.6"
pynfe = "^0.4.4"
dropbox = "^11.36.2"
pip = "^23.3.2"
install = "^1.3.5"
pyairtable = "^2.2.1"

[tool.poetry.dev-dependencies]
pytest = ">=7.0.0"


[tool.poetry.group.dev.dependencies]
black = ">=23.10.1"
ruff = ">=0.1.4"
pylance = ">=0.8.11"
pytest = ">=7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
